refactor: @ alias 적용, 절대 경로 수정
