+++ src/refactoring/hooks/useCoupon.ts
@@ -1,5 +1,5 @@
+import { Coupon } from "@/types";
 import { useCallback, useState } from "react";
-import { Coupon } from "../../types.ts";
 
 export const useCoupons = (initialCoupons: Coupon[]) => {
   const [coupons, setCoupons] = useState<Coupon[]>(initialCoupons);
