+++ src/refactoring/services/permission.ts
@@ -1,4 +1,4 @@
-import { PERMISSION_TYPE } from "../../types";
+import { PERMISSION_TYPE } from "@/types";
 
 // 갑자기 궁금한 부분은.. 서버와 통신하는 함수들은
 export const getUserPermission = async () => {
