+++ vite.config.ts
@@ -1,10 +1,16 @@
-import { defineConfig as defineTestConfig, mergeConfig } from "vitest/config";
-import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react-swc";
+import path from "path";
+import { defineConfig } from "vite";
+import { defineConfig as defineTestConfig, mergeConfig } from "vitest/config";
 
 export default mergeConfig(
   defineConfig({
     plugins: [react()],
+    resolve: {
+      alias: {
+        "@": path.resolve(__dirname, "./src"),
+      },
+    },
   }),
   defineTestConfig({
     test: {
